pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mosparo_api_client"
version = "1.1.0"
description = "Python API Client to communicate with mosparo."
readme = "README.md"
authors = [{ name = "mosparo Core Developers", email = "info@mosparo.io" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
]
keywords = ["mosparo", "api-client", "spam-protection", "accessibility", "captcha"]
dependencies = [
    "requests>=2.31.0"
]
requires-python = ">=3.5"

[project.optional-dependencies]
dev = ["pytest-runner", "requests-mock", "pip-tools"]

[project.urls]
Website = "https://mosparo.io"
GitHub = "https://github.com/mosparo/python-api-client"
